Add Verse Highlighting
* Verse highlighting released for testing.  Enable it by checking enableVerseHighlighting and restarting app
* Toggle the markers on and off in Layout / Toggle Highlight Markers
* Highlight markers show up only in verse view to the left of the verses
* There are currently 3 markers (yellow, blue, and underline)
* There are some problems with viewing certain Bibles in formatted chapter view
* This needs speed improvements.  Not sure why it has lag responding to commands.
