+++ BibleBooks.py
@@ -683,6 +683,8 @@ class BibleBooks:
         "LUK.": "42",
         "Luke": "42",
         "jn": "43",
+        "jo": "43",
+        "joh": "43",
         "jhn": "43",
         "john": "43",
         "Jhn.": "43",
