+++ UniqueBibleAppVersion.txt
@@ -1,1 +1,1 @@
-19.3
+19.4
