+++ main.py
@@ -327,6 +327,12 @@ if not hasattr(config, "logCommands"):
 # Migrate Bible name from Verses table to Details table
 if not hasattr(config, "migrateDatabaseBibleNameToDetailsTable"):
     config.migrateDatabaseBibleNameToDetailsTable = True 
+# Verse highlighting functionality
+if not hasattr(config, "enableVerseHighlighting"):
+    config.enableVerseHighlighting = False
+# Show verse highlight markers
+if not hasattr(config, "showHighlightMarkers"):
+    config.showHighlightMarkers = True
 
 # Setup logging
 logger = logging.getLogger('uba')
@@ -512,6 +518,7 @@ def saveDataOnExit():
         ("enableCopyHtmlCommand", config.enableCopyHtmlCommand),
         ("forceGenerateHtml", config.forceGenerateHtml),
         ("logCommands", config.logCommands),
+        ("enableVerseHighlighting", config.enableVerseHighlighting),
         ("migrateDatabaseBibleNameToDetailsTable", config.migrateDatabaseBibleNameToDetailsTable)
     )
     with open("config.py", "w", encoding="utf-8") as fileObj:
