+++ patches.txt
@@ -2,14 +2,14 @@
 (18.6, "folder", "htmlResources/fonts")
 (18.3, "folder", "htmlResources/css")
 (18.3, "folder", "htmlResources/js")
+(19.4, "file", "Languages.py")
+(19.4, "file", "TextCommandParser.py")
+(19.4, "file", "gui/RemoteControl.py")
+(19.4, "file", "patches.txt")
+(19.4, "file", "latest_changes.txt")
+(19.4, "file", "UniqueBibleAppVersion.txt")
 (19.3, "file", "gui/MainWindow.py")
-(19.3, "file", "gui/RemoteControl.py")
-(19.3, "file", "Languages.py")
-(19.3, "file", "patches.txt")
-(19.3, "file", "latest_changes.txt")
-(19.3, "file", "UniqueBibleAppVersion.txt")
 (19.2, "file", "BiblesSqlite.py")
-(19.2, "file", "TextCommandParser.py")
 (19.2, "file", "ThirdParty.py")
 (19.2, "file", "gui/CentralWidget.py")
 (19.2, "file", "install.sh")
